chore: add the possibility to disable the auto updater

+++ node/src/main/java/eu/cloudnetservice/node/module/updater/ModuleUpdater.java
@@ -32,29 +32,31 @@ public final class ModuleUpdater implements Updater<ModuleUpdaterContext> {
   private static final Logger LOGGER = LogManager.logger(ModuleUpdater.class);
 
   @Override
-  public void executeUpdates(@NonNull ModuleUpdaterContext context) {
-    FileUtil.walkFileTree(DefaultModuleProvider.DEFAULT_MODULE_DIR, ($, file) -> {
-      // check if we already know an associated module
-      var moduleName = context.moduleNames().get(file.toAbsolutePath());
-      if (moduleName != null) {
-        // check if the module is an official module which gets updates from remote
-        context.modules().findByName(moduleName).ifPresent(moduleEntry -> {
-          // validate using the current checksum if the file is up-to-date
-          var currentChecksum = ChecksumUtil.fileShaSum(file);
-          if (!moduleEntry.sha3256().equals(currentChecksum)) {
-            // there is an update available - download it!
-            Unirest
-              .get(moduleEntry.url(context.updaterRepo(), context.updaterBranch()))
-              .asFile(file.toString(), StandardCopyOption.REPLACE_EXISTING);
-            // validate the checksum now
-            var newModuleChecksum = ChecksumUtil.fileShaSum(file);
-            if (!moduleEntry.sha3256().equals(newModuleChecksum)) {
-              LOGGER.warning(I18n.trans("cloudnet-load-modules-invalid-checksum", moduleName));
-              FileUtil.delete(file);
+  public void executeUpdates(@NonNull ModuleUpdaterContext context, boolean onlyIfRequired) {
+    if (!onlyIfRequired) {
+      FileUtil.walkFileTree(DefaultModuleProvider.DEFAULT_MODULE_DIR, ($, file) -> {
+        // check if we already know an associated module
+        var moduleName = context.moduleNames().get(file.toAbsolutePath());
+        if (moduleName != null) {
+          // check if the module is an official module which gets updates from remote
+          context.modules().findByName(moduleName).ifPresent(moduleEntry -> {
+            // validate using the current checksum if the file is up-to-date
+            var currentChecksum = ChecksumUtil.fileShaSum(file);
+            if (!moduleEntry.sha3256().equals(currentChecksum)) {
+              // there is an update available - download it!
+              Unirest
+                .get(moduleEntry.url(context.updaterRepo(), context.updaterBranch()))
+                .asFile(file.toString(), StandardCopyOption.REPLACE_EXISTING);
+              // validate the checksum now
+              var newModuleChecksum = ChecksumUtil.fileShaSum(file);
+              if (!moduleEntry.sha3256().equals(newModuleChecksum)) {
+                LOGGER.warning(I18n.trans("cloudnet-load-modules-invalid-checksum", moduleName));
+                FileUtil.delete(file);
+              }
             }
-          }
-        });
-      }
-    }, false, "*.{jar,war,zip}");
+          });
+        }
+      }, false, "*.{jar,war,zip}");
+    }
   }
 }
